% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metabo_data_analyzer.R
\name{Metabo_data_analyzer}
\alias{Metabo_data_analyzer}
\title{Metabo_data_analyzer}
\usage{
Metabo_data_analyzer(datapath, pathresult = NULL, author = "LNA", use_IPO = FALSE, ret_corvar=FALSE, intvalpar="into", center=NULL, normalization=TRUE, NomRapport = NULL, grouping_mode, no_conditioning_QC = "ON", list_conQC= "NULL", RT_max)
}
\arguments{
\item{datapath}{the path source where is the data to analyze}

\item{pathresult}{the path where to store the analyzed data}

\item{author}{the name of project author. Default to Luca}

\item{use_IPO}{either use IPO or not to determine the best wcms parameters for peak picking. (Not available yet)}

\item{ret_corvar}{chose the method to align the samples between "density" and "obiwarp"}

\item{intvalpar}{what kind of peak measurement should the algorithm use? default to IPO}

\item{centerQC}{should the obiwarp method centered to a specific QC sample}

\item{normalization}{should the sample batch normalized?}

\item{NomRapport}{what is the name of Rapport? default to a name obtained from the datapath}

\item{grouping_mode}{what kind of grouping mode should CAMERA use to determine the pseudospectra? chose between EIC_corr or area_corr}

\item{no_conditioning_QC}{should be conditioning QC erased from the normalization process? Default set to yes}

\item{list_conQC}{list of conditioning QC. If null, the algorithm will try to determine what are the conditioning QCs}

\item{RT_max}{What is the limit of RT to consider a peak included in the peaklist?}
}
\value{
peaky
}
\description{
The function transforms you raw mzML data files in a deconvoluted peaklist eliminating also the the redundant informations
}
\examples{
peaky <- Metabo_data_analyzer(datapath = datapathy, grouping_mode = "area_corr", RT_max = 17)

}
\author{
Luca Narduzzi "nardluca@gmail.com"
}
